Remove code duplication and add tests
